공통 BackButton 컴포넌트 생성 - 뒤로가기 버튼 중복 제거
