+++ src/components/common/index.ts
@@ -4,6 +4,7 @@ export { default as Select } from "./Select";
 export { default as Toggle } from "./Toggle";
 export { default as Card } from "./Card";
 export { default as CollapsibleCard } from "./CollapsibleCard";
+export { default as BackButton } from "./BackButton";
 export { default as Badge } from "./Badge";
 export { default as DataRow } from "./DataRow";
 export { default as Skeleton } from "./Skeleton";
